+++ core/src/main/java/com/facebook/ktfmt/format/Formatter.kt
@@ -42,6 +42,8 @@ import org.jetbrains.kotlin.psi.psiUtil.startOffset
 
 object Formatter {
 
+  @JvmField val META_FORMAT = FormattingOptions()
+
   @JvmField
   val GOOGLE_FORMAT =
       FormattingOptions(
@@ -66,7 +68,7 @@ object Formatter {
    */
   @JvmStatic
   @Throws(FormatterException::class, ParseError::class)
-  fun format(code: String): String = format(FormattingOptions(), code)
+  fun format(code: String): String = format(META_FORMAT, code)
 
   /**
    * format formats the Kotlin code given in 'code' with 'removeUnusedImports' and returns it as a
@@ -75,7 +77,7 @@ object Formatter {
   @JvmStatic
   @Throws(FormatterException::class, ParseError::class)
   fun format(code: String, removeUnusedImports: Boolean): String =
-      format(FormattingOptions(removeUnusedImports = removeUnusedImports), code)
+      format(META_FORMAT.copy(removeUnusedImports = removeUnusedImports), code)
 
   /**
    * format formats the Kotlin code given in 'code' with the 'maxWidth' and returns it as a string.